requirements.txt
colorama==0.4.6
contourpy==1.3.2
cupy-cuda12x==13.6.0
cycler==0.12.1
fastrlock==0.8.3
fonttools==4.61.1
h5py==3.14.0
kiwisolver==1.4.9
matplotlib==3.10.8
mkl-service==2.4.0
nptyping==2.5.0
numba==0.61.2
numpy==1.26.4
packaging==26.0
pillow==12.1.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.17.0
tqdm==4.67.3
